This is the base HTML and CSS for my portfolio site.
This includes most of the base bones for the site. I still have a lot
more in-depth stuff to do here, but it should be a sold start. I need
to research more about positioning things, because I'm starting to
doubt that manipulting margins is the best way to go. :-/